+++ app/Http/Controllers/ProductoController.php
@@ -3,50 +3,56 @@
 namespace App\Http\Controllers;
 
 use App\Models\Categoria;
+use App\Models\MotivoMovimiento;
+use App\Models\Movimiento;
 use App\Models\Producto;
+use App\Models\Registro;
+use App\Models\TipoMovimiento;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
+use Illuminate\Validation\Rule;
 
 class ProductoController extends Controller
 {
     public function index(Request $request)
     {
-        $query = Producto::with('categoria');
-
-        if ($request->has('search')) {
-            $query->where('nombre_producto', 'like', '%'.$request->search.'%');
+        if (request()->user('admin')->cannot('viewAny', Producto::class)) {
+            return view('admin.productos.index'); // salir sin enviar datos
         }
 
-        // Ordenamiento
-        switch ($request->ordering) {
-            case 'recientes':
-                $query->orderBy('created_at', 'desc');
-                break;
-            case 'nombre_asc':
-                $query->orderBy('nombre_producto', 'asc');
-                break;
-            case 'nombre_desc':
-                $query->orderBy('nombre_producto', 'desc');
-                break;
-            case 'precio_asc':
-                $query->orderBy('precio', 'asc');
-                break;
-            case 'precio_desc':
-                $query->orderBy('precio', 'desc');
-                break;
-            default:
-                $query->orderBy('created_at', 'desc');
-                break;
+        $query = Producto::query()->with('categoria');
+
+        // Filtro por categoría
+        if ($categoria_id = request('categoria')) {
+            $query->where('categoria_id', $categoria_id);
         }
 
-        // Paginación con parámetros persistentes
-        $productos = $query->paginate(10)->appends($request->only(['search', 'ordering']));
+        // Filtro por estado de stock
+        if ($estado = request('estado_stock')) {
+            $query->when($estado === 'agotado', fn ($q) => $q->where('stock_actual', 0))
+                ->when($estado === 'bajo', fn ($q) => $q->whereBetween('stock_actual', [1, 5]))
+                ->when($estado === 'normal', fn ($q) => $q->where('stock_actual', '>', 5));
+        }
 
-        return view('admin.productos.index', compact('productos'));
+        // Filtro por estado activo/inactivo
+        if (! is_null(request('activo'))) {
+            $query->where('activo', request('activo') === '1');
+        }
+
+        // realiza búsqueda y retorna datos paginados
+        $productos = Producto::busqueda($request, $query);
+        $categorias = Categoria::all();
+
+        return view('admin.productos.index', compact('productos', 'categorias'));
     }
 
     public function create()
     {
+        if (request()->user('admin')->cannot('create', Producto::class)) {
+            abort(403);
+        }
+
         $categorias = Categoria::all();
 
         return view('admin.productos.create', [
@@ -56,15 +62,18 @@ class ProductoController extends Controller
 
     public function store(Request $request)
     {
-        // TODO: Implementar las siguientes validaciones
-        // - Producto ya existe
+
+        if (request()->user('admin')->cannot('create', Producto::class)) {
+            abort(403);
+        }
 
         $request->validate([
             'nombre_producto' => ['required', 'max:250', 'unique:productos'],
-            'categoria'       => ['required'],
+            'categoria'       => ['nullable'],
+            'destacado'       => [],
             'descripcion'     => [],
-            'stock_actual'    => ['gte:0', 'max:9'],
-            'precio'          => ['required', 'string', 'max:12'],
+            'stock_actual'    => ['required', 'string'],
+            'precio'          => ['required', 'string'],
             'imagen'          => ['image', 'mimes:jpeg,png,jpg,webp', 'max:2048'],
         ]);
 
@@ -72,39 +81,86 @@ class ProductoController extends Controller
         $precioRaw = $request->input('precio'); // "$12.345"
         $precio = (int) str_replace(['$', '.'], '', $precioRaw); // 12345
 
+        // Limpiamos el string de stock_actual y lo dejamos como entero
+        $stock = (int) str_replace('.', '', $request->input('stock_actual'));
+
         // para la imágen
         if ($request->imagen) {
-            $url_imagen = '/images/productos/'.time().'.'.$request->imagen->extension();
-            $request->imagen->move(public_path('images/productos'), $url_imagen);
+            $imagen_url = $request->imagen->store('images/productos');
         }
 
-        Producto::create([
+        $producto = Producto::create([
             'nombre_producto' => request('nombre_producto'),
-            'categoria_id'    => request('categoria'),
+            'slug'            => Str::slug(request('nombre_producto')),
+            'categoria_id'    => request('categoria') ?? null,
+            'destacado'       => request()->has('destacado'),
             'descripcion'     => request('descripcion'),
-            'stock_actual'    => request('stock_actual'),
+            'stock_actual'    => $stock,
             'precio'          => $precio,
-            'imagen_url'      => $url_imagen ?? null,
+            'imagen_url'      => $imagen_url ?? null,
         ]);
 
-        session()->flash('success_create', '¡Producto creado exitosamente!');
+        // registrar acción del admin
+        Registro::registrar_accion($producto, 'Crea nuevo producto');
+
+        session()->flash('success', '¡Producto creado exitosamente!');
 
         return redirect('/admin/productos/');
     }
 
-    /**
-     * Display the specified resource.
-     * Usado para vista detallada en pagina de producto
-     * NOTA: A futuro implementar slug en vez de id para mejorar visualizacion
-     * de la url y posicionamiento
-     * ej url con id =   producto/15
-     * ej url con slug = producto/polera-oversize-blanca
-     */
-    public function show($id)
+    public function show(Producto $producto)
     {
-        $producto = Producto::where('id', $id)->firstOrFail();
+        $tipos = TipoMovimiento::all();
+        $motivos = MotivoMovimiento::all();
 
-        return view('producto.show', compact('producto'));
+        // Validamos los filtros
+        request()->validate([
+            'tipo'    => 'nullable|integer|exists:tipo_movimientos,id',
+            'motivo'  => 'nullable|integer|exists:motivo_movimientos,id',
+            'ordenar' => 'nullable|in:fecha_asc,fecha_desc,cantidad_asc,cantidad_desc',
+        ]);
+
+        $movimientos = Movimiento::query()
+            ->with('producto')
+            ->where('producto_id', $producto->id);
+
+        // Aplicar filtro por tipo
+        if ($tipo_id = request('tipo')) {
+            $movimientos->where('tipo_movimiento_id', $tipo_id);
+        }
+
+        // Aplicar filtro por motivo
+        if ($motivo_id = request('motivo')) {
+            $movimientos->where('motivo_movimiento_id', $motivo_id);
+        }
+
+        // Ordenar por cantidad o fecha
+        switch (request('ordenar')) {
+            case 'fecha_asc':
+                $movimientos->orderBy('created_at', 'asc');
+                break;
+            case 'fecha_desc':
+                $movimientos->orderBy('created_at', 'desc');
+                break;
+            case 'cantidad_asc':
+                $movimientos->orderBy('cantidad', 'asc');
+                break;
+            case 'cantidad_desc':
+                $movimientos->orderBy('cantidad', 'desc');
+                break;
+            default:
+                $movimientos->orderBy('created_at', 'desc');
+                break;
+        }
+
+        $movimientos = $movimientos->paginate(20);
+
+        return view('admin.productos.show', compact(
+            'producto',
+            'movimientos',
+            'tipos',
+            'motivos',
+        ));
     }
 
     /**
@@ -112,6 +168,10 @@ class ProductoController extends Controller
      */
     public function edit(Producto $producto)
     {
+        if (request()->user('admin')->cannot('update', Producto::class)) {
+            abort(403);
+        }
+
         $categorias = Categoria::all();
 
         return view('admin.productos.edit', compact('producto', 'categorias'));
@@ -122,12 +182,16 @@ class ProductoController extends Controller
      */
     public function update(Request $request, Producto $producto)
     {
+        if (request()->user('admin')->cannot('update', Producto::class)) {
+            abort(403);
+        }
+
         $request->validate([
-            'nombre_producto' => ['required', 'max:250', 'unique:productos'],
-            'categoria'       => ['required'],
+            'nombre_producto' => ['required', 'max:250', Rule::unique('productos')->ignore($producto->id)],
+            'categoria'       => ['nullable'],
             'descripcion'     => [],
-            'stock_actual'    => ['nullable', 'gte:0', 'max:9'],
-            'precio'          => ['required', 'string', 'max:12'],  // el maximo es 12 por los puntos y $
+            'stock_actual'    => ['required', 'string'],
+            'precio'          => ['required', 'string'],
             'imagen'          => ['image', 'mimes:jpeg,png,jpg,webp', 'max:2048'],
         ]);
 
@@ -135,29 +199,34 @@ class ProductoController extends Controller
         $precioRaw = $request->input('precio'); // "$12.345"
         $precio = (int) str_replace(['$', '.'], '', $precioRaw); // 12345
 
+        // Limpiamos el string de stock_actual y lo dejamos como entero
+        $stock = (int) str_replace('.', '', $request->input('stock_actual'));
+
         // para la imágen
         if ($request->imagen) {
-            $url_imagen = '/images/productos/'.time().'.'.$request->imagen->extension();
-            $request->imagen->move(public_path('images/productos'), $url_imagen);
+            $imagen_url = $request->imagen->store('images/productos');
 
-            // TODO: eliminar imagen anterior
-            /* if ($producto->imagen_url && File::exists($producto->imagen_url)) { */
-            /*     File::delete($producto->imagen_url); */
-            /* } */
+            if ($producto->imagen_url && Storage::exists($producto->imagen_url)) {
+                Storage::delete($producto->imagen_url);
+            }
         }
 
         $producto->update([
             'nombre_producto' => $request->input('nombre_producto'),
-            'categoria_id'    => $request->input('categoria'),
+            'slug'            => Str::slug($request->input('nombre_producto')),
+            'categoria_id'    => $request->input('categoria') ?? null,
+            'destacado'       => $request->has('destacado'),
             'descripcion'     => $request->input('descripcion'),
-            'stock_actual'    => $request->input('stock_actual'),
+            'stock_actual'    => $stock,
             'precio'          => $precio,
-            'imagen_url'      => $url_imagen ?? $producto->imagen_url,
+            'imagen_url'      => $imagen_url ?? $producto->imagen_url,
         ]);
 
-        session()->flash('success_update', '¡Producto actualizado exitosamente!');
+        Registro::registrar_accion($producto, 'Edita un producto');
 
-        return redirect()->route('productos.index');
+        session()->flash('success', '¡Producto actualizado exitosamente!');
+
+        return redirect()->route('productos.show', $producto);
     }
 
     /**
@@ -165,10 +234,35 @@ class ProductoController extends Controller
      */
     public function destroy(Producto $producto)
     {
+        if (request()->user('admin')->cannot('delete', Producto::class)) {
+            abort(403);
+        }
+
         $producto->delete();
 
-        session()->flash('success_delete', '¡Producto eliminado exitosamente!');
+        Registro::registrar_accion($producto, 'Elimina un producto');
+
+        session()->flash('success', '¡Producto eliminado exitosamente!');
 
         return redirect()->route('productos.index');
+    }
+
+    public function disable(Producto $producto)
+    {
+        if (request()->user('admin')->cannot('update', Producto::class)) {
+            abort(403);
+        }
+
+        if ($producto->activo) {
+            $producto->update(['activo' => false]);
+            session()->flash('success', 'Producto desactivado');
+            Registro::registrar_accion($producto, 'Desactiva un producto');
+        } else {
+            $producto->update(['activo' => true]);
+            session()->flash('success', 'Producto reactivado');
+            Registro::registrar_accion($producto, 'Reactiva un producto');
+        }
+
+        return redirect()->back();
     }
 }
